buffer.c: bufferDataAvail() function added. It tells if data is available in the buffer or not.
modeminit.c:
states are added to the code. Based on state modem is initalised.
This gives the benefit to directly jumping to any of the state.
Initialisation is done till getting IP address only.
Error handling is not yet implememted.
sendwait() does not contain response as its parameter. response is based on the state.